backend/doncoin/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
bitarray==3.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==1.0.2
cssselect2==0.8.0
cytoolz==1.1.0
Django==5.2.9
django-cors-headers==4.9.0
django-filer==3.4.1
django-filter==25.2
django-polymorphic==4.1.0
djangorestframework==3.16.1
easy-thumbnails==2.10.1
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Faker==38.2.0
frozenlist==1.8.0
hexbytes==0.3.1
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lru-dict==1.2.0
lxml==6.0.2
multidict==6.7.0
parsimonious==0.10.0
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
psycopg2==2.9.11
py-ecc==8.0.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
pytz==2025.2
pyunormalize==17.0.0
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
rlp==4.1.0
rpds-py==0.30.0
sqlparse==0.5.4
svglib==1.5.1
tinycss2==1.5.1
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
web3==6.15.1
webencodings==0.5.1
websockets==15.0.1
yarl==1.22.0
